Address review feedback on branch management
- Move branch tests from BaseIcebergConnectorTest to
  BaseConnectorTest with SUPPORTS_BRANCH behavior guard,
  so any connector supporting branches gets test coverage.
  Keep testBranchVisibleInRefsTable in BaseIcebergConnectorTest
  since $refs is Iceberg-specific.
- Add SUPPORTS_BRANCH(false) to TestingConnectorBehavior.
- Add testMetadataDeleteFromBranch for full-table deletes.
- Use INVALID_ARGUMENTS instead of NOT_SUPPORTED when
  rejecting DROP BRANCH main, since branch dropping is
  supported — the argument is invalid.
- Add verify() null checks on branch ref lookups after
  commit in finishInsert statistics and summary resolution.
