+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/IcebergMetadata.java
@@ -1750,7 +1750,11 @@ public class IcebergMetadata
         }
         else if (!computedStatistics.isEmpty()) {
             long newSnapshotId = table.branch()
-                    .map(branch -> icebergTable.refs().get(branch).snapshotId())
+                    .map(branch -> {
+                        SnapshotRef ref = icebergTable.refs().get(branch);
+                        verify(ref != null, "Branch ref not found after commit: %s", branch);
+                        return ref.snapshotId();
+                    })
                     .orElseGet(() -> icebergTable.currentSnapshot().snapshotId());
 
             CollectedStatistics collectedStatistics = processComputedTableStatistics(icebergTable, computedStatistics);
@@ -1768,7 +1772,11 @@ public class IcebergMetadata
         transaction = null;
 
         Snapshot resultSnapshot = table.branch()
-                .map(branch -> icebergTable.snapshot(icebergTable.refs().get(branch).snapshotId()))
+                .map(branch -> {
+                    SnapshotRef ref = icebergTable.refs().get(branch);
+                    verify(ref != null, "Branch ref not found after commit: %s", branch);
+                    return icebergTable.snapshot(ref.snapshotId());
+                })
                 .orElseGet(icebergTable::currentSnapshot);
         Map<String, String> summary = resultSnapshot != null ? resultSnapshot.summary() : null;
         if (summary == null) {
@@ -2556,7 +2564,7 @@ public class IcebergMetadata
         IcebergTableHandle table = checkValidTableHandle(tableHandle);
 
         if (branch.equals(SnapshotRef.MAIN_BRANCH)) {
-            throw new TrinoException(NOT_SUPPORTED, "Cannot drop the main branch");
+            throw new TrinoException(INVALID_ARGUMENTS, "Cannot drop the main branch");
         }
 
         BaseTable icebergTable = catalog.loadTable(session, table.getSchemaTableName());
